requirements.txt
requests==2.19.1
redis==2.10.6
pymongo==3.7.1
mongoengine==0.15.3
lxml==4.2.3
selenium==3.13.0
django==2.0.7
pyexecjs==1.5.1
python-json-logger==0.1.9
pillow==5.2.0
pytesseract==0.2.4